Update template for Terraform 1.1
Why these changes are being introduced:
We are making use of the cloud{} block with workspace tags to make it
easier for an infra dev to be able to link cli-driven speculative
plans to different workspaces associated with a single repository.

How this addresses that need:
* Replace `backend "remote"` with `cloud`
* Replace workspace name with workspace tags
* Update version constraint for terraform to 1.1 or newer
* Fix a few minor typos

Side effects of this change:
None.