Test Shared Workflows
Why these changes are being introduced:
EngX started experimenting with shared workflows for their Ruby code.
Based on the fact that we have the same three workflows in all of our
Terraform repositories build off this template, it makes sense to
consolidate those workflows into one shared repo and then call the
shared copies from the Terraform repo.

The changes here are dependent on the changes to the .github repository
in our organization:
https://github.com/MITLibraries/.github

Right now, the new "caller" workflow in this template references
the shared (e.g. "called") workflows in the feature branch as we are
still testing. Once that feature branch is merged to `main` this repo
will need one more commit to change the {ref} at the end of the
`uses:` line in the "caller" workflow.

How this addresses that need:
* Create a new workflow that calls the three shared workflows.
* Remove the local workflows since they are redundant

Side effects of this change:
None