Update references to called workflows
After the PR & merge to main in the .github repository, the
caller workflow in this repo needed to be updated to pull from the
main branch, not the feature branch. This update also includes some
minor updates to the naming inside the caller workflow for clarity
in the Actions tab in the repo in GitHub.

Side effects:
None