repo update and maintenance
Why these changes are being introduced:

This commit add new standardized workflows to this repository. It expands the
files and documentaion and brings them up to our new standards. This should
significantly reduce the amount of work it takes for us to deploy new apps
off of this template and reduce errors and missing code/docs while doing so.
Finally, this commit updates the version of terraform and pre-commit action
dependency programs.

How this addresses that need:
 - Add pre-commit hooks, files, and code
 - Update terraform version to v1.10
 - Update all workflow app versions
 - SECURITY: restrict GH workflows to readonly
 - Add CODEOWNERS file
 - Add set/delete initial SSM param scripts
 - README.md documentation updates
 - Add placeholder .terraform.lock.hcl

Side effects of this change:

None

Changes to be committed:
      new file:   .github/CODEOWNERS
      modified:   .github/workflows/tf-shared-workflows.yml
      new file:   .pre-commit-config.yaml
      new file:   .terraform-docs.yaml
      new file:   .terraform.lock.hcl
      modified:   README.md
      modified:   files/README.md
      new file:   files/delete_initial_ssm_param.sh
      new file:   files/set_initial_ssm_param.sh
      modified:   versions.tf